*: Move from our fork of sysinfo to upstream and get test suite passing on AMD cpus (#7682) (#8939)
