+++ components/engine_rocks/Cargo.toml
@@ -16,6 +16,7 @@ failpoints = ["fail/failpoints"]
 encryption = { path = "../encryption" }
 engine = { path = "../engine" }
 engine_traits = { path = "../engine_traits" }
+num_cpus = "1"
 prometheus = { version = "0.8", features = ["nightly", "push"] }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
