+++ components/tikv_util/Cargo.toml
@@ -29,6 +29,7 @@ hex = "0.4"
 lazy_static = "1.3"
 libc = "0.2"
 log = { version = "0.4", features = ["max_level_trace", "release_max_level_debug"] }
+num_cpus = "1"
 openssl = "0.10"
 prometheus = { version = "0.8", features = ["nightly", "push"] }
 quick-error = "1.2.2"
@@ -40,7 +41,7 @@ slog-async = "2.3"
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
 slog-json = "2.3"
 slog-term = "2.4"
-sysinfo = { git = "https://github.com/tikv/sysinfo.git", branch = "tikv"}
+sysinfo = "0.14"
 tikv_alloc = { path = "../tikv_alloc" }
 time = "0.1"
 tokio-core = "0.1"
