+++ components/tikv_util/src/sys/mod.rs
@@ -22,7 +22,7 @@ pub mod sys_quota {
         }
 
         pub fn cpu_cores_quota(&self) -> f64 {
-            let cpu_num = sysinfo::get_logical_cores() as f64;
+            let cpu_num = num_cpus::get() as f64;
             match self.cgroup.cpu_cores_quota() {
                 Some(cgroup_quota) if cgroup_quota > 0.0 && cgroup_quota < cpu_num => cgroup_quota,
                 _ => cpu_num,
@@ -64,7 +64,7 @@ pub mod sys_quota {
         }
 
         pub fn cpu_cores_quota(&self) -> f64 {
-            sysinfo::get_logical_cores() as f64
+            num_cpus::get() as f64
         }
 
         pub fn memory_limit_in_bytes(&self) -> u64 {
